requirements.txt
django
djangorestframework
environs
markdown
django-filter
django-environ
psycopg2
djangorestframework_simplejwt
